Modify and write methods, create test class, edit annotation syntax
Modify initializeVectorMap() to include restraints on parameters ‘rule’
and ‘seed’ and write the displayVectorMap() method in the Automata
class. Reorder packages in alphabetical order and fix syntax for method
annotations in the GeneratorTest class. Fix syntax for method
annotations in the VectorTest class. Write test class AutomatonTest
which tests initializeVector() in the Automaton class.