Modify and add functionality to UI and include manifest file
AutomatonImage: modify the paint component method so it no longer relies on a Graphics instance. Simplify code for the image-to-map method and write a method to resize an image.
MainApp: Add label for displaying automata. Fix tick marks and snap behavior for both sliders. Set default value for rule to 30. Include seed type options in the combo box. Modify program logic for the display button which will display a cellular automaton in the display panel.
Manifest: Include manifest file in order to run project as a .jar file.