Add event listeners to display button and size slider
MainApp: add event listener to display button so that a cellular automaton displays. Remove the next and previous buttons as well as the speed slider. Instead of displaying each generation at a time as initially planned, display all generations at once. Double maximum value of the speed slider and modify the tick spacing. Add event listener so that the current size text field updates every time the slider is dragged.